compute_lb

+++ src/dist_table.rs
@@ -7,14 +7,14 @@ use crate::{graph::Graph, instance::Instance};
 #[derive(Debug, Clone)]
 pub struct DistTable<'a> {
     pub num_vertices: usize,
-    pub table: Vec<Vec<isize>>,
+    pub table: Vec<Vec<i32>>,
     pub open_queues: Vec<VecDeque<usize>>,
     graph: &'a Graph,
 }
 
 impl<'a> DistTable<'a> {
     pub fn new_lazy(instance: &'a Instance) -> Self {
-        let max_dist = instance.graph.vertices.len() as isize;
+        let max_dist = instance.graph.vertices.len() as i32;
         let mut table = vec![vec![max_dist; instance.graph.vertices.len()]; instance.num_agents];
         let mut open_queues = vec![VecDeque::new(); instance.num_agents];
         for i in 0..instance.num_agents {
@@ -33,7 +33,7 @@ impl<'a> DistTable<'a> {
 
     pub fn new_parallel(instance: &'a Instance) -> Self {
         let num_vertices = instance.graph.vertices.len();
-        let max_dist = num_vertices as isize;
+        let max_dist = num_vertices as i32;
         let mut table = vec![vec![max_dist; num_vertices]; instance.num_agents];
 
         table
@@ -63,9 +63,9 @@ impl<'a> DistTable<'a> {
     }
 
     pub fn get(&mut self, agent_id: usize, vertex_id: usize) -> isize {
-        let max_dist = self.num_vertices as isize;
+        let max_dist = self.num_vertices as i32;
         if self.table[agent_id][vertex_id] < max_dist {
-            return self.table[agent_id][vertex_id];
+            return self.table[agent_id][vertex_id] as isize;
         }
 
         while let Some(current_vertex_id) = self.open_queues[agent_id].pop_front() {
@@ -82,7 +82,7 @@ impl<'a> DistTable<'a> {
             }
 
             if current_vertex_id == vertex_id {
-                return current_dist;
+                return current_dist as isize;
             }
         }
